woop: rename to useWoopHandling()
