+++ client/signup/steps/woocommerce-install/hooks/use-woop-handling/index.ts
